Remove the enc package
Move the remaining files to types/
