+++ types/blob_leaf_codec.go
@@ -1,4 +1,4 @@
-package enc
+package types
 
 import (
 	"bytes"
