+++ types/codec_test.go
@@ -1,4 +1,4 @@
-package enc
+package types
 
 import (
 	"bytes"
@@ -15,11 +15,11 @@ func TestEncode(t *testing.T) {
 	// Encoding details for each codec are tested elsewhere.
 	// Here we just want to make sure codecs are selected correctly.
 	dst := &bytes.Buffer{}
-	Encode(dst, bytes.NewReader([]byte{0x00, 0x01, 0x02}))
+	encode(dst, bytes.NewReader([]byte{0x00, 0x01, 0x02}))
 	assert.Equal([]byte{'b', ' ', 0x00, 0x01, 0x02}, dst.Bytes())
 
 	dst.Reset()
-	Encode(dst, []interface{}{42})
+	encode(dst, []interface{}{42})
 	assert.Equal("t [42]\n", string(dst.Bytes()))
 }
 
@@ -27,18 +27,18 @@ func TestInvalidDecode(t *testing.T) {
 	assert := assert.New(t)
 
 	d.IsUsageError(assert, func() {
-		Decode(bytes.NewReader([]byte{}))
+		decode(bytes.NewReader([]byte{}))
 	})
 
 	d.IsUsageError(assert, func() {
-		Decode(bytes.NewReader([]byte{0xff}))
+		decode(bytes.NewReader([]byte{0xff}))
 	})
 }
 
 func TestSelectBlobDecoder(t *testing.T) {
 	assert := assert.New(t)
 
-	decoded := Decode(bytes.NewReader([]byte{'b', ' ', 0x2B}))
+	decoded := decode(bytes.NewReader([]byte{'b', ' ', 0x2B}))
 	out := &bytes.Buffer{}
 	_, err := io.Copy(out, decoded.(io.Reader))
 	assert.NoError(err)
@@ -46,6 +46,6 @@ func TestSelectBlobDecoder(t *testing.T) {
 }
 
 func TestSelectTypedDecoder(t *testing.T) {
-	v := Decode(bytes.NewBufferString(`t [42]`))
+	v := decode(bytes.NewBufferString(`t [42]`))
 	assert.Equal(t, []interface{}{float64(42)}, v)
 }
